Pose Estimation Sagmaker/Model Requirment : Dependencies/requirements.txt
torch==1.7.0
torchvision==0.8.1
Pillow
dlib
opencv-python

